Refactor and finalize
