+++ app/styles/components/departure.css
@@ -3,10 +3,11 @@
 
 
 .departure-box {
-  /*border: 1px solid color-black;*/
+  border: 1px solid color-black;
   border-radius: 3px;
   padding: 0 10px;
   margin: 15px 0;
+  background: white;
   box-shadow: color-light-black 0px 1px 6px, color-light-black 0px 1px 4px
 }
