+++ app/styles/containers/app-container.css
@@ -1,9 +1,66 @@
 @value colors: "../layout/colors.css";
-@value color-blue from colors;
+@value color-blue, color-background, color-background-transparent, color-dark-gray from colors;
+@value fonts: "../layout/fonts.css";
+@value main-font from fonts;
+
+body {
+  font-family: main-font;
+  margin: 0;
+}
 
 .app {
   /*color: color-blue;*/
-  min-height: 100%;
+  min-height: 100vh;
   height: 100%;
   width: 100%;
+}
+
+
+.app:before {
+  content: ' ';
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  background-image: linear-gradient(color-background-transparent, color-background-transparent), url('../../assets/images/coachella.jpg');
+  background-size: cover;
+  will-change: transform;
+  z-index: -1;
+}
+
+.app:after {
+  /*content: '';
+  background-color: color-background-transparent;
+  opacity: 0.5;
+  position: absolute;
+  height: 100%;
+  width: 100%;*/
+}
+
+.center-loading {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.header {
+  text-align: center;
+  margin-top: 2em;
+  margin-bottom: 2em;
+}
+
+.header img {
+  max-width: 100%;
+  text-align: center;
+}
+
+button.fetch-button {
+  border: 1px solid color-dark-gray;
+  color: color-dark-gray;
+  background: transparent;
+  border-radius: 0;
+  padding: 20px;
+  font-size: 2rem;
 }
