+++ app/styles/layout/fonts.css
@@ -1,0 +1,1 @@
+@value main-font: 'Catamaran', sans-serif;
