+++ package.json
@@ -35,6 +35,7 @@
     "file-loader": "^0.9.0",
     "postcss-loader": "^1.1.1",
     "style-loader": "^0.13.1",
+    "url-loader": "^0.5.7",
     "webpack": "^1.3.0",
     "webpack-dev-server": "^1.16.2"
   },
